requirements.txt
langchain==0.3.23
langchain-chroma==0.2.2
langchain-google-genai==2.0.0
chromadb==0.4.24
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pysqlite3-binary==0.5.1
numpy<2.0.0
